Extract UserSettingsListener to DRY setting stores
